backend/requirements.txt
fastapi==0.95.2
uvicorn==0.23.2
openai==0.27.8
pinecone-client==3.0.0
beautifulsoup4==4.12.2
httpx==0.24.1
langchain==0.0.90
python-dotenv==1.0.0
requests==2.31.0
